document endpoint removed | send message validated
