--- MangoAPI.BusinessLogic/ApiCommands/Documents/UploadDocumentCommandValidator.cs
@@ -1,17 +1,0 @@
-﻿using FluentValidation;
-using MangoAPI.BusinessLogic.Pipelines;
-
-namespace MangoAPI.BusinessLogic.ApiCommands.Documents;
-
-public class UploadDocumentCommandValidator : AbstractValidator<UploadDocumentCommand>
-{
-    public UploadDocumentCommandValidator()
-    {
-        RuleFor(x => x.UserId).NotEmpty();
-        RuleFor(x => x.ContentType).NotEmpty();
-
-        RuleFor(x => x.FormFile)
-            .NotNull()
-            .SetValidator(new CommonFileValidator());
-    }
-}
