+++ MangoAPI.BusinessLogic/ApiCommands/Users/UpdateProfilePictureCommandValidator.cs
@@ -1,5 +1,4 @@
 ﻿using FluentValidation;
-using MangoAPI.BusinessLogic.Pipelines;
 
 namespace MangoAPI.BusinessLogic.ApiCommands.Users;
